rework: updateAvatar moved from userVC to userVM, add setAvatarImage on StatisticCell
