+++ FakeNFT/Presentation/StatisticsFlow/User Screen/UserViewModel.swift
@@ -6,6 +6,7 @@
 //
 
 import UIKit
+import Kingfisher
 
 final class UserViewModel: UserViewModelProtocol {
     // MARK: - Observables
@@ -19,4 +20,20 @@ final class UserViewModel: UserViewModelProtocol {
     }
     
     // MARK: - Methods
+    
+    func updateAvatar(for imageView: UIImageView) {
+        guard let url = URL(string: user.avatar) else {
+            return
+        }
+        
+        let cache = ImageCache.default
+        cache.diskStorage.config.expiration = .seconds(1)
+        
+        let processor = RoundCornerImageProcessor(cornerRadius: 35, backgroundColor: .clear)
+        imageView.kf.indicatorType = .activity
+        imageView.kf.setImage(with: url,
+                                 placeholder: nil,
+                                 options: [.processor(processor),
+                                           .cacheSerializer(FormatIndicatedCacheSerializer.png)])
+    }
 }
