+++ FakeNFT/Protocols/UserViewModelProtocol.swift
@@ -5,10 +5,11 @@
 //  Created by Aleksandr Garipov on 17.08.2023.
 //
 
-import Foundation
+import UIKit
 
 protocol UserViewModelProtocol {
     var user: UserModel { get }
     var userObservable: Observable<UserModel>  { get }
+    func updateAvatar(for imageView: UIImageView)
     init(user: UserModel)
 }
